refactor: separate send email logic
