+++ backend/app/schemas/send_email.py
@@ -6,3 +6,12 @@ class SendEmailRequest(BaseModel):
     subject: str
     body: str
     copy_emails: list[EmailStr] = []
+
+class SendOpportunityEmailRequest(SendEmailRequest):
+    email: EmailStr
+    subject: str
+    body: str
+    copy_emails: list[EmailStr] = []
+    opportunity_id: int
+    file_name: str
+    file_base64: str
